chore: move test codec into test utils
