+++ crates/p2p/src/test_utils.rs
@@ -1,2 +1,3 @@
 pub mod main_loop;
 pub mod peer_aware;
+pub mod sync;
